+++ src/mol-gl/shader/chunks/assign-marker-varying.glsl.ts
@@ -1,5 +1,5 @@
 export const assign_marker_varying = `
-#if defined(dRenderVariant_color) || defined(dRenderVariant_marking)
+#if defined(dNeedsMarker)
     #if defined(dMarkerType_instance)
         vMarker = readFromTexture(tMarker, aInstance, uMarkerTexDim).a;
     #elif defined(dMarkerType_groupInstance)